requirements.txt
feedparser
gspread
google-auth
beautifulsoup4
requests
